force instance of frames for binding
Made a bunch of changes to force individual instance creation on frame change. This solves the issues with bindings failing to work. Thank you to the (paid) developer who helped devise the solution. (credit can be provided if you send me a message. I'll add a tag if they agree to it but at the time of the writing they have not responded and I'm looking forward to getting this published) Implementation here is slightly different than the provided solution as we don't need to pass anything to change_frame() as there is an existing variable value to track the frames.

Binding is now working as expected.